tools/requirements.txt
ImageHash==4.2.0
numpy==1.19.4
Pillow==8.0.1
pymemcache==3.4.0
PyWavelets==1.1.1
scipy==1.5.4
six==1.15.0
